api-server/pyproject.toml
[project]
name = "api-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "beanie>=1.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.14",
    "fastapi-jwt-auth>=0.5.0",
    "motor>=3.7.1",
    "python-dotenv>=1.1.1",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]
